% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{interval}
\usage{
interval(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A scaled version of x which varies between 0 and 1
}
\description{
interval
}
\examples{
a <- 1:10
interval(a)
}
