% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{reverse}
\usage{
reverse(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector
}
\description{
Reverses a vector so that the last element comes first and the first last
}
\examples{
x <- 1:10
reverse(x)
}
