% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{center}
\usage{
center(x)
}
\arguments{
\item{x}{a vector of numeric or integer values}
}
\value{
a vector of numeric values
}
\description{
center returns a centered copy of a vector of values. To do this, it
subtracts the mean of the vector from the value.
}
\examples{
a <- 50:100
center(a)
}

