% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse}
\usage{
reverse(x)
}
\arguments{
\item{x}{}
}
\value{
a vector
}
\description{
Reverse returns a copy of a vector whose elements are in
the reverse order. The end.
}
\details{
See also \code{\link{rev}}
}
\examples{
reverse(1:10)
}

