% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse a vector}
\usage{
reverse(x)
}
\arguments{
\item{x}{}
}
\value{
a vector
}
\description{
Reverse returns the copy of a vector whose elements are in the reverse order.
}
\examples{
x <- 1:10
reverse(x)
}

